Add files via upload
The ATM Program in C is written in C programming language which provides an ease to read and comprehend the instructions used. This program for using ATM machine is built on the concept of handling an account individually.

It can be defined as actually simple code structure of ATM transaction process to be understood by a user. For implementing this project, we may have to use function but in the meantime for easy coding, we may have to switch cause statement.

From this ATM program in C, we can even use the mini-program for checking the total balance, depositing the amount, and withdrawing the amount from the account definitely since it is not time overwhelming.